Create a Members and Member class
